requirements.txt
discord.py==2.3.2
bs4==0.0.2
selenium==4.17.2
python-dotenv==1.0.1